backend/requirements.txt
fastapi
uvicorn
python-dotenv
pyyaml
langchain
langchain-community
langchain-openai
langchain-google-genai
langgraph
pydantic
